refactor: secure user profiles and fix database schema errors - Refactored JWT and middleware for better context management. - Fixed price type mismatch (int -> float64) in product models. - Improved database migration reporting.

+++ database/dbmodel/products.go
@@ -8,7 +8,7 @@ type Product struct {
 	gorm.Model
 	NameProduct string `json:"name_product"`
 	Description string `json:"description"`
-	Price int `json:"price"`
+	Price float64 `json:"price"`
 	IdTheme int
 	Theme Theme `gorm:"foreignKey:IdTheme"` 
 	IdCategorie int
